Move `option.js`
